+++ internal/platform/seed/seed.go
@@ -25,7 +25,7 @@ type Options struct {
 var rolePermissions = map[string][]string{
 	"admin": {
 		"user.read", "user.create", "user.update",
-		"post.read", "post.create", "post.update", "post.publish",
+		"post.read", "post.create", "post.update", "post.publish", "post.archive",
 		"tag.create", "tag.update", "tag.delete",
 		"settings.read", "settings.update",
 		"media.create", "media.delete",
@@ -34,7 +34,7 @@ var rolePermissions = map[string][]string{
 	},
 	"editor": {
 		"user.read",
-		"post.read", "post.create", "post.update", "post.publish",
+		"post.read", "post.create", "post.update", "post.publish", "post.archive",
 		"tag.create", "tag.update", "tag.delete",
 		"media.create", "media.delete",
 		"comments.moderate",